#1730, Add support to silent running for process in ZK UI, reference to issue: https://github.com/adempiere/adempiere/issues/1730
