+++ zkwebui/WEB-INF/src/org/adempiere/webui/apps/ProcessPanel.java
@@ -627,23 +627,11 @@ public class ProcessPanel extends ProcessController implements SmallViewEditable
 	@Override
 	public void afterInit() {
 		//	BR [ 265 ]
-		if (!hasParameters()) {
-			if (getShowHelp() != null 
-					&& getShowHelp().equals("N")) {
-				setAutoStart(true);    // don't ask first click
-				// anyway show resulting window
-			}
-		}
-		// Check if the process is a silent one
-		if(getShowHelp() != null 
-				&& getShowHelp().equals("S")) {
-			setAutoStart(true);
-		}
+		validateAutoStart();
 		//	
 		if(!isAutoStart()) {
 			loadQuerySaved();
 		} else if(parent.getParentProcess() == null) {
-			process();
 			Events.postEvent("onClick", bOK, null);
 		}
 	}
